korpus/backend/requirements.txt
argcomplete==2.0.0
arrow==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
blessed==1.19.1
Brotli==1.0.9
cachetools==5.2.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.0.12
colorclass==2.2.2
compressed-rtf==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssselect2==0.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
detect==2020.12.3
Django==4.1
django-concurrency==2.4
django-filter==22.1
django-picklefield==3.1
django-q==1.3.9
django-spa==0.3.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docx==0.2.4
docx2txt==0.8
drf-yasg==1.21.3
easygui==0.98.3
ebcdic==1.1.1
elastic-transport==8.1.2
elasticsearch==8.4.1
et-xmlfile==1.1.0
extract-msg==0.36.2
fonttools==4.37.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
gunicorn==20.1.0
html5lib==1.1
htmldocx==0.0.6
httplib2==0.21.0
idna==3.3
IMAPClient==2.3.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kwargs==1.0.1
lark-parser==0.12.0
lxml==4.9.1
MarkupSafe==2.1.1
msoffcrypto-tool==5.0.0
mysql-connector-python==8.0.30
mysqlclient==2.1.1
oauthlib==3.2.2
olefile==0.46
oletools==0.60.1
only==2020.12.3
openpyxl==3.0.10
packaging==21.3
pcodedmp==1.2.6
pdfminer.six==20220524
Pillow==9.2.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydyf==0.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pyphen==0.12.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
RTFDE==0.0.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
textract==1.6.5
tinycss2==1.1.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
Wand==0.6.10
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
zopfli==0.2.1
